upadte

+++ main/arobota2/script/publish_STM_omni.py
@@ -40,9 +40,9 @@ class STM_Connect():
 
     def Update_Speed(self, msg):
         self.wheel = msg.data.split(',')
-        self._right_wheel_power = int(self.wheel[0])/800
-        self._center_wheel_power = int(self.wheel[1])/800
-        self._left_wheel_power = int(self.wheel[2])/800
+        self._right_wheel_power = float(self.wheel[0])/800
+        self._center_wheel_power = float(self.wheel[1])/800
+        self._left_wheel_power = float(self.wheel[2])/800
         # print ("M2A"+str(int(float(self.wheel[0])*1000000))+"B"+str(int(float(self.wheel[1])*1000000))+"\r\n")
         # rospy.loginfo(msg.data)
         #power_message = "M0"+"A"+str(int(self._right_wheel_power)/500)+"B"+str(int(self._center_wheel_power)/500)+"C"+str(int(self._left_wheel_power)/500)+"\r\n"
